+++ src/format.rs
@@ -14,12 +14,47 @@ use crate::{
     rng::Rng,
 };
 
+#[derive(Debug, Default, Clone, PartialEq, Eq)]
 #[allow(dead_code)]
 enum Padding {
+    #[default]
     None,
     Left,
     Right,
     Both,
+}
+
+#[derive(Debug, Default, Clone, PartialEq, Eq)]
+#[allow(dead_code)]
+enum LineEnd {
+    #[default]
+    Unchanged,
+    ForceAdd,
+    ForceRemove,
+}
+
+#[derive(Debug, Default, Clone)]
+struct FmtOptions {
+    padding: Padding,
+    line_end: LineEnd,
+}
+
+impl FmtOptions {
+    fn with_padding(mut self, padding: Padding) -> Self {
+        self.padding = padding;
+        self
+    }
+    fn with_line_end(mut self, line_end: LineEnd) -> Self {
+        self.line_end = line_end;
+        self
+    }
+}
+
+impl From<&FormattingOptions> for FmtOptions {
+    fn from(_: &FormattingOptions) -> Self {
+        // TODO: Modify if necessary
+        Self::default()
+    }
 }
 
 pub(crate) trait Format {
@@ -29,17 +64,18 @@ pub(crate) trait Format {
 impl Format for &ParolLs {
     fn format(&self, options: &FormattingOptions) -> Vec<TextEdit> {
         let range = <&ParolLs as Into<Rng>>::into(*self).0;
-        let new_text = self.txt(options);
+        let fmt_options = options.into();
+        let new_text = self.txt(&fmt_options);
         vec![TextEdit { range, new_text }]
     }
 }
 
 trait Fmt {
-    fn txt(&self, options: &FormattingOptions) -> String;
+    fn txt(&self, options: &FmtOptions) -> String;
 }
 
 impl Fmt for ASTControl {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         match self {
             ASTControl::ASTControl0(_) => "^".to_string(),
             ASTControl::ASTControl1(ut) => ut.user_type_declaration.txt(options),
@@ -47,7 +83,7 @@ impl Fmt for ASTControl {
     }
 }
 impl Fmt for Alternation {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         self.alternation_list
             .iter()
             .map(|a| a.txt(options))
@@ -56,16 +92,20 @@ impl Fmt for Alternation {
     }
 }
 impl Fmt for AlternationList {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
+        let comment_options = options
+            .clone()
+            .with_line_end(LineEnd::ForceRemove)
+            .with_padding(Padding::Left);
         format!(
             "{}{}",
             self.factor.txt(options),
-            handle_comments(&*self.comments, Padding::Left, options)
+            handle_comments(&*self.comments, &comment_options)
         )
     }
 }
 impl Fmt for Alternations {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{} {}",
             self.alternation.txt(options),
@@ -79,27 +119,28 @@ impl Fmt for Alternations {
     }
 }
 impl Fmt for AlternationsList {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
+        let comment_options = options.clone().with_padding(Padding::Both);
         format!(
             "\n    {} {}{}",
             self.or,
-            handle_comments(&*self.comments, Padding::Both, options),
+            handle_comments(&*self.comments, &comment_options),
             self.alternation.txt(options)
         )
     }
 }
 impl Fmt for BlockComment {
-    fn txt(&self, _options: &FormattingOptions) -> String {
+    fn txt(&self, _options: &FmtOptions) -> String {
         self.block_comment.text().to_string()
     }
 }
 impl Fmt for CommentsList {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         self.comments_list_group.txt(options)
     }
 }
 impl Fmt for CommentsListGroup {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         match self {
             CommentsListGroup::CommentsListGroup0(l) => l.line_comment.txt(options),
             CommentsListGroup::CommentsListGroup1(b) => b.block_comment.txt(options),
@@ -107,24 +148,25 @@ impl Fmt for CommentsListGroup {
     }
 }
 impl Fmt for CutOperator {
-    fn txt(&self, _options: &FormattingOptions) -> String {
+    fn txt(&self, _options: &FmtOptions) -> String {
         self.cut_operator.text().to_string()
     }
 }
 impl Fmt for Declaration {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
+        let comment_options = options.clone().with_padding(Padding::Left);
         match self {
             Declaration::Declaration0(title) => format!(
                 "{} {}{}\n",
                 title.percent_title,
                 title.string.txt(options),
-                handle_comments(&*title.comments, Padding::Left, options),
+                handle_comments(&*title.comments, &comment_options),
             ),
             Declaration::Declaration1(comment) => format!(
                 "{} {}{}\n",
                 comment.percent_comment,
                 comment.string.txt(options),
-                handle_comments(&*comment.comments, Padding::Left, options),
+                handle_comments(&*comment.comments, &comment_options),
             ),
             Declaration::Declaration2(user_type) => format!(
                 "{} {} {} {}{}\n",
@@ -132,7 +174,7 @@ impl Fmt for Declaration {
                 user_type.identifier.txt(options),
                 user_type.equ,
                 user_type.user_type_name.txt(options),
-                handle_comments(&*user_type.comments, Padding::Left, options),
+                handle_comments(&*user_type.comments, &comment_options),
             ),
             Declaration::Declaration3(scanner_directives) => {
                 handle_scanner_directives(&*scanner_directives.scanner_directives, options)
@@ -142,12 +184,12 @@ impl Fmt for Declaration {
 }
 
 impl Fmt for DoubleColon {
-    fn txt(&self, _options: &FormattingOptions) -> String {
+    fn txt(&self, _options: &FmtOptions) -> String {
         self.double_colon.text().to_string()
     }
 }
 impl Fmt for Factor {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         match self {
             Factor::Factor0(g) => g.group.txt(options),
             Factor::Factor1(r) => r.repeat.txt(options),
@@ -157,7 +199,7 @@ impl Fmt for Factor {
     }
 }
 impl Fmt for GrammarDefinition {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{}{}{}",
             self.percent_percent,
@@ -172,12 +214,12 @@ impl Fmt for GrammarDefinition {
     }
 }
 impl Fmt for GrammarDefinitionList {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         self.production.txt(options)
     }
 }
 impl Fmt for Group {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{} {}{}\n",
             self.l_paren,
@@ -187,17 +229,17 @@ impl Fmt for Group {
     }
 }
 impl Fmt for Identifier {
-    fn txt(&self, _options: &FormattingOptions) -> String {
+    fn txt(&self, _options: &FmtOptions) -> String {
         self.identifier.text().to_string()
     }
 }
 impl Fmt for LineComment {
-    fn txt(&self, _options: &FormattingOptions) -> String {
+    fn txt(&self, _options: &FmtOptions) -> String {
         self.line_comment.text().to_string()
     }
 }
 impl Fmt for NonTerminal {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{}{}",
             self.identifier.identifier,
@@ -208,12 +250,12 @@ impl Fmt for NonTerminal {
     }
 }
 impl Fmt for NonTerminalOpt {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         self.a_s_t_control.txt(options)
     }
 }
 impl Fmt for Optional {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{} {}{}",
             self.l_bracket,
@@ -223,7 +265,7 @@ impl Fmt for Optional {
     }
 }
 impl Fmt for ParolLs {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{}\n{}",
             self.prolog.txt(options),
@@ -232,7 +274,7 @@ impl Fmt for ParolLs {
     }
 }
 impl Fmt for Production {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "\n{}{}\n    {}",
             self.production_l_h_s.txt(options),
@@ -242,18 +284,20 @@ impl Fmt for Production {
     }
 }
 impl Fmt for ProductionLHS {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
+        let comment_options_both = options.clone().with_padding(Padding::Both);
+        let comment_options_right = options.clone().with_padding(Padding::Right);
         format!(
             "\n{}{}{}\n    {} ",
-            handle_comments(&*self.comments, Padding::Right, options),
+            handle_comments(&*self.comments, &comment_options_right),
             self.identifier.identifier,
-            handle_comments(&*self.comments0, Padding::Both, options),
+            handle_comments(&*self.comments0, &comment_options_both),
             self.colon,
         )
     }
 }
 impl Fmt for Prolog {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{}{}{}",
             self.start_declaration.txt(options),
@@ -269,17 +313,17 @@ impl Fmt for Prolog {
     }
 }
 impl Fmt for PrologList {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         self.declaration.txt(options)
     }
 }
 impl Fmt for PrologList0 {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         self.scanner_state.txt(options)
     }
 }
 impl Fmt for Repeat {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{} {}{}",
             self.l_brace,
@@ -289,12 +333,12 @@ impl Fmt for Repeat {
     }
 }
 impl Fmt for ScannerDirectives {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         handle_scanner_directives(self, options)
     }
 }
 impl Fmt for ScannerState {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{} {} {}\n{} {}\n",
             self.percent_scanner,
@@ -312,12 +356,12 @@ impl Fmt for ScannerState {
     }
 }
 impl Fmt for ScannerStateList {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         handle_scanner_directives(&self.scanner_directives, options)
     }
 }
 impl Fmt for ScannerSwitch {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         match self {
             ScannerSwitch::ScannerSwitch0(sc) => format!(
                 "{}{}{}{}",
@@ -342,12 +386,12 @@ impl Fmt for ScannerSwitch {
     }
 }
 impl Fmt for ScannerSwitchOpt {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         self.identifier.txt(options)
     }
 }
 impl Fmt for SimpleToken {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{}{}",
             self.string.string,
@@ -358,23 +402,25 @@ impl Fmt for SimpleToken {
     }
 }
 impl Fmt for SimpleTokenOpt {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         self.a_s_t_control.txt(options)
     }
 }
 impl Fmt for StartDeclaration {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
+        let comment_options_left = options.clone().with_padding(Padding::Left);
+        let comment_options_right = options.clone().with_padding(Padding::Right);
         format!(
             "{}{} {}{}\n",
-            handle_comments(&*self.comments, Padding::Right, options),
+            handle_comments(&*self.comments, &comment_options_right),
             self.percent_start,
             self.identifier.txt(options),
-            handle_comments(&*self.comments0, Padding::Left, options),
+            handle_comments(&*self.comments0, &comment_options_left),
         )
     }
 }
 impl Fmt for StateList {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{}{}",
             self.identifier.txt(options),
@@ -388,22 +434,22 @@ impl Fmt for StateList {
     }
 }
 impl Fmt for StateListList {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!("{} {}", self.comma, self.identifier.txt(options),)
     }
 }
 impl Fmt for crate::parol_ls_grammar_trait::String {
-    fn txt(&self, _options: &FormattingOptions) -> String {
+    fn txt(&self, _options: &FmtOptions) -> String {
         self.string.text().to_string()
     }
 }
 impl Fmt for Symbol {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         handle_symbol(self, options)
     }
 }
 impl Fmt for TokenWithStates {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{}{}{}{}{}",
             self.l_t,
@@ -417,17 +463,17 @@ impl Fmt for TokenWithStates {
     }
 }
 impl Fmt for TokenWithStatesOpt {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         self.a_s_t_control.txt(options)
     }
 }
 impl Fmt for UserTypeDeclaration {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!("{} {}", self.colon, self.user_type_name.txt(options),)
     }
 }
 impl Fmt for UserTypeName {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{}{}",
             self.identifier.txt(options),
@@ -441,7 +487,7 @@ impl Fmt for UserTypeName {
     }
 }
 impl Fmt for UserTypeNameList {
-    fn txt(&self, options: &FormattingOptions) -> String {
+    fn txt(&self, options: &FmtOptions) -> String {
         format!(
             "{}{}",
             self.double_colon.double_colon,
@@ -452,36 +498,39 @@ impl Fmt for UserTypeNameList {
 
 fn handle_scanner_directives(
     scanner_directives: &ScannerDirectives,
-    options: &FormattingOptions,
+    options: &FmtOptions,
 ) -> String {
+    let comment_options = options.clone().with_padding(Padding::Left);
     match scanner_directives {
         ScannerDirectives::ScannerDirectives0(l) => format!(
             "{} {}{}\n",
             l.percent_line_underscore_comment,
             l.string.txt(options),
-            handle_comments(&*l.comments, Padding::Left, options),
+            handle_comments(&*l.comments, &comment_options),
         ),
         ScannerDirectives::ScannerDirectives1(b) => format!(
             "{} {} {}{}\n",
             b.percent_block_underscore_comment,
             b.string.txt(options),
             b.string0.txt(options),
-            handle_comments(&*b.comments, Padding::Left, options),
+            handle_comments(&*b.comments, &comment_options),
         ),
+
         ScannerDirectives::ScannerDirectives2(n) => format!(
             "{}{}\n",
             n.percent_auto_underscore_newline_underscore_off,
-            handle_comments(&*n.comments, Padding::Left, options),
+            handle_comments(&*n.comments, &comment_options),
         ),
+
         ScannerDirectives::ScannerDirectives3(w) => format!(
             "{}{}\n",
             w.percent_auto_underscore_ws_underscore_off,
-            handle_comments(&*w.comments, Padding::Left, options),
+            handle_comments(&*w.comments, &comment_options),
         ),
     }
 }
 
-fn handle_symbol(symbol: &Symbol, options: &FormattingOptions) -> String {
+fn handle_symbol(symbol: &Symbol, options: &FmtOptions) -> String {
     match symbol {
         Symbol::Symbol0(n) => n.non_terminal.txt(options),
         Symbol::Symbol1(t) => t.simple_token.txt(options),
@@ -490,7 +539,7 @@ fn handle_symbol(symbol: &Symbol, options: &FormattingOptions) -> String {
     }
 }
 
-fn handle_comments(comments: &Comments, padding: Padding, options: &FormattingOptions) -> String {
+fn handle_comments(comments: &Comments, options: &FmtOptions) -> String {
     let comments = comments
         .comments_list
         .iter()
@@ -501,11 +550,22 @@ fn handle_comments(comments: &Comments, padding: Padding, options: &FormattingOp
     if comments.is_empty() {
         comments
     } else {
-        match padding {
-            Padding::None => comments,
-            Padding::Left => format!(" {}", comments),
-            Padding::Right => format!("{} ", comments),
-            Padding::Both => format!(" {} ", comments),
-        }
+        apply_formatting(comments, options)
+    }
+}
+
+fn apply_formatting(line: String, options: &FmtOptions) -> String {
+    let line = match options.line_end {
+        LineEnd::Unchanged => line,
+        LineEnd::ForceAdd => line + "\n",
+        LineEnd::ForceRemove => line
+            .trim_end_matches(|c| c == '\r' || c == '\n')
+            .to_string(),
+    };
+    match options.padding {
+        Padding::None => line,
+        Padding::Left => format!(" {}", line),
+        Padding::Right => format!("{} ", line),
+        Padding::Both => format!(" {} ", line),
     }
 }